requirements.txt
backports.functools-lru-cache==1.4
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
pandas==0.20.3
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn-pandas==1.5.0
subprocess32==3.2.7
